+++ NEWS.md
@@ -1,3 +1,13 @@
+# mau 0.4.0
+* Update function read_utilities, now script with utilities functions it is not created for each read.
+
+# mau 0.3.0
+* New standard for naming functions
+  
+# mau 0.2.0
+* Adding new functions for computations and aggregation of rating `colley_rating`, `od_rating`, 
+  `borda_count`
+
 # mau 0.1.2
 
 * `Bar.Plot` Improving bar plots.